Fix number search response, and make country field not required
This fixes DEVX-1881 which covers the optional country field on the
endpoint for listing owned numbers (we made it required because we just
reused an existing required field - this refactors that).

We also corrected the response fields in the search for available
numbers endpoint - this was a copy of the owned numbers one, which
wasn't accurate.