Making search_pattern description a little clearer
Thinking that this description should perhaps change to note that the country code is inclusive for value 0 - for US developers, in particular, we tend to think in terms of our area codes (321, 201, 305, etc. . .) And as I stumbled on this one I think others might be confused by it as well.